+++ src/main/java/com/pi/tobeeb/Repositorys/ConsultationFileRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ConsultationFileRepository extends JpaRepository<ConsultationFile,Long> {
-    Optional<ConsultationFile> findByAppointment_IdAppointment(Long idAppointment);
+    Optional<ConsultationFile> findByAppointment_IdAppointment(int idAppointment);
 
     List<ConsultationFile> findAllByAppointment_Patient(User user);
 }
